Interrupt enabled and set up for pin 15
We are using a Change Notification to interrupt on a change. When the
button on pin 15 is pushed it will go high. If we enter that routine we
set the buttonFlag and return to the main loop. if the buttonFlag is set
we will go ahead and display the message.