Redefine angleThreshold as extern const
angleThreshold was redefined multiple times within main.c when it is in
fact a constant value representing the amount of handle motion that
should be considered significant (more than just noise). Therefore,
angleThreshold was removed from main, defined as "extern" in
IWPUtilities.h, and assigned the value of 0.08 in IWPUtilities.c. This
makes angleThreshold a constant global variable that can be assigned a
value once, and then accessed anywhere inside IWP_Firmware.