Replaced hard-coded 10ms delay, volumeDelay, & 1ms Leakage Rate Loop delay w/ upstrokeInterval
upstrokeInterval is (at the time of this revision) a 10ms delay designed
to be a constant amount of elapsed time between polling the
accelerometer for a new pump handle position. Throughout the code
inconsistent hard-coded delays of 10ms and 1ms were in use as the
interval between checking the accelerometer for handle movement. This
situation has been rectified.  A volumeDelay that was equal to the
upstrokeInterval was also replaced with upstrokeInterval.