volumeEvent adjustment for leakRate
volumeEvent should adjust the volume of water pumped to account for the
amount of water leaking out of the pump while pumping based on the leak
rate which is reported in L/s:

volumeEvent -= (leakRate * extractionDuration / upstrokeInterval);

extractionDuration is a loop counter variable and every time through the
loop upstrokeInterval number of ms has elapsed. Therefore, it looks like
you're taking the volume event and subtracting out ([L/s] * [unitless
counter]) / [10ms]. I believe this is not going to be in Liters.
Therefore this calculation has been adjusted for correct dimensional
analysis.