ADC 10bit calculation -> 12bit calculation
It was determined in the code review 12/3/16 that the ADC of the PIC was
updated from 10bit to 12bit. Therefore, the adjustment factor to divide
the range of the ADC into positive and negative values by suptracting
the value of (2^10)/2 from the ADC output no longer makes sense. The new
adjustment factor is (2^12)/2. Additionally, the adjustment factor was
renamed from adjustmentFactor -> signedNumAdjustADC for clarity.