Definition of PI
the new constant variable PI was given a value in both IPWUtilities.c
and .h.  To be consistent with other variables, .h now has PI declared
as an external variable but its value is assigned in .c  This allows a
clean build, the other way had the linker confused since PI was defined
two places.