Saving Info to EEPROM and Report at Noon not midnight
Added code to LeakRate, PrimingTime and Volume calculations to save
values expected to be reported later to EEProm.
Wrote a noonMessage function intended to replace the midnightMessage.
Removed the code which sends an init message since we now expect to be
in the initilization routine each time the PIC comes back to life when
running on only Solar power.