Turning on WPS and Shutdown at night if battery is low
Moved the 555 On/Off outside of the readWaterSensor function.  Now the
555 is turned on when the handle starts to move and off when volume and
leak have been calculated.  This gives the 555 time to stableize before
trying to read it.
Added a check at the start of the Main loop so that the typical routine
is only run if the battery is > 3.0V or if it is not, the time is
between 6AM and 6PM