Debug Message Function
In IWPUtilities I added a sendDebugMessage function.  It is passed a
comment that you want and a float variable.  These are sent to the
Serial Out pins usually used to talk to the FONA.  Since they don't use
the AT protocol, the FONA ignores them and you can read the messages
using any tool (DIGILENT) that watches a UART protocol to send messages
when debugging.