Clear WDT while trying to connect to the Network
The code tries to connect to the network for 5min.  Since this exceeds
the nominal 2min WDT timeout, I added the ClearWatchDogTimer command
within the 1 minute attempt loop.  If we get to this command, we are not
hung (and thus want the WDT) we are knowingly spending time outside of
the Main loop which will clear the WDT.