Leak Rate Loop changed
The new leak rate loop only declares that pumping has restarted if the
handle has moved more than angleThresholdSmall 15 times in a row.  This
is the same reasoning used in the Volume calculation to declare that the
handle has stopped moving (movement < angleThresholdSmall 15 times) .
The old code never encountered leak condition 2 which means that there
was still water after the declared limit for watching it to leak out
(currently 3 sec).  Now we break out of the leak rate and use the old if
someone starts pumping (condition 1), declare the leak rate to be zero
if we have waited 3 sec and there is still water (condition 2), or
calculate a leak rate based upon how long it took for the water to go
down if it goes down in less than 3 sec but no one has started pumping
(case 3).