WDT and Data Saved in EEProm
When EEProm unused values are initialized after sending the noon
message, I put 0.01 in the volume locations.  This will allow us to see
that the value reported was never written to at all.  If it says 0, it
might be a valid report of 0.

Added a function to save the volume information from the previous time
slot to EEProm.  This is called in the main loop rather than saving this
information in the volume calculation loop as it was before.  Now if the
time slot changes, the volume calculated for the previous bin will be
saved even if no one is pumping during the current time slot.

Added a DEBUG routine DebugReadEEProm which will read and display the
values saved in EEProm to the serial port so they can be read

To save power, 24mA down to 9mA, I turn off the FONA board if we are
sleeping at night.  I don't just call turnOffSIM, I check first to see
if it is on.

Added the code to call SaveVolumeToEEProm if we are in a new volume
pumped time bin

Added a call to Clear the Watch Dog Time in the Handle Idle loop since
we are here as long as no one is pumping and that is likley to be more
than the 130sec time out for the WDT

Added a call to Clear the Watch Dog Time in the pumping loop.  If
someone pumps for more than the 130sec without stopping, we don't want
to time out for the WDT

Removed reading the current hour from inside the volume loop since it
was just read when pumping started.  We will assign volume pumped to the
time slot when the pumping began.

Removed the saving of volume data to EEProm from the Volume loop since
it was made part of the main outside loop so that it still gets done
even if no one is pumping.