Moved Sleep routine inside of Handle movement
The code stays in the handle movement routine unless someone pumps.  We
need to go to sleep while waiting for handle movement if it is
appropriate and we need to update EEProm with past volume information if
it is time.  So I moved both of these functions to the start of the
handle movement loop.  I tested the sleep function and it properly goes
in and comes out.