Initial Clearing of EEProm
The first time we run the initialization function, we want to clear
EEProm locations.  This includes EEProm#20 which we use for Restart
Status.  If Restart Status = 0, we know that the board was once alive.
We check this in subsequent calls to initialization which will happen if
we die because of a bad battery overnight.  If it is found to be 0, we
know to reset maxleakrate and maxpriming time to the value saved from
the previous afternoon.