Changes to the Priming Loop and Leak Rate
We used to wait 7 seconds if someone stopped pumping and considered that
time still spent trying to Prime.  Now there are three ways we get out
of the priming loop.
1.) Water shows up
2.) Someone stops trying for 1 sec.  We calculate the Priming time as
though we got water in 1sec.  As far as we know, unless we have a longer
one already, priming takes "at least" one second.  We also set a new
flag called never_primed.
3.) They tried for 7 seconds but never got any water.  This is the
longest we wait for priming.  We calculate the priming time as though we
got water after 7 seconds.  As far as we know it takes "at least" 7
seconds to prime the pump.  We also set the never_primed flag.

In the Leak Rate loop, if the never_primed flag is set, we are in a new
case, case 4.  In this case there was never any water so we can't
calculate a leak.  When this is true, we assume the leak rate was zero.