Added switch to turn printing of debug messages on and off
printing the debug messages to the serial port takes some time and may
mess up normal operation a little.  So we don't want them to print
unless we are working on the system.  I added a flag called
print_debug_messages.  This is set to 0 during initialization.  If it is
0, when the sendDebuMessage function is called, it will not do anything.
We would leave things in this state normally.

If you want messages printed, set the flag to 1 at the top of the Main
loop