Handle Moving Detection
Changed the debug report statement so it sends the hour message rather
than the noon message.  This is what was intended all along.

Changed the detection of handle movement to be 5 degrees from the handle
angle read at the start of the loop.  Before we measured handle movement
many times and accumulated the delta.  This was way too sensitive to
noise.  This new approach just keeps compairing the current handle angle
to the one measured when entering the loop.  Once they are 5 degrees
different, we declare that the handle is moving.

NOTE:  the 5 degrees is hard coded and should be made a constant.  Not
sure why I don't just do that but I'm feeling skiddish