Leak Rate Change and small Priming Change
The timOutStatus++ was only being incremented if the handle moved more
than a "noise" amount.  It should be incremented each time through the
loop.

In the Leak Rate, I found that we get the 427.507 leak rate if we get to
the LeakRate routine and there is no water as soon as we get there.
This is a wierd condition that we need to research.  For now, I added a
check for the presence of water as soon as we enter the LeakRate
routine.  If it is not there, I declare a leakCondition of 4 which means
we don't have enough good data to calculate a new leak rate.  This is
the same as when the pump never primed in the first place.  Actually,
this can probibly replace that test but I'm not messing with it.  I also
changed the case 4 so that it would use the previous leak rate rather
than 0.  If we don't know enough to calculate a new value, we don't know
enough to use a value of 0 in our calculation