Fix Possible Hang Conditions with FONA, other changes see description
turnOffSIM, turnONSIM waited until success.  Now I just try to turn
on/off for the proscribed 2sec and added a return variable that
indicates if it worked or not.
tryToConnectToNetwork kept turning FONA on and off.  If the network is
not there after 7 times looking for it.  The routine now returns a
variable indicating that the network was not found.
In the main loop, we try to send the noon message as long as it is noon
and we have not yet been able to send it because we could not connect to
the network.  If it turns 1PM, we give up, flush the previous day's data
and start over.
In the Connected to Network function I changed the timing to use TMR1
rather than a dealy loop and some unecessary calculations.  If the
lights are not flashing at all, we break out of the routine rather than
hang. and return a variable saying we were not able to connect.
ResetMsgVariables now also writes 0.01 to unused EEPROM so it can be
called and done at 1PM if noon message was not sent
in noonMessage I stopped using volumeRAM locations when converting to
strings and used a local dummy variable.  I think the RAM locations were
getting left with yesterday's data if no new volume readings happened in
those hours.
In SaveVolumetoEEPROM I cleared the volume RAM locations after saving
their data to EEPROM so we don't just add the next day's volume to the
previous.
Added a constant handleMovementThreshold in IWPutilites to define how
many degrees the handle must move from rest before we say the handle is
moving