Fixed I2C to avoid having the code hang
Changed configI2c BaudRateGen BRG setting to give desired 100khz clock
the old setting gave 48khz.
Changed turnOffClockOscilator, getHourI2C and setTime functions so that
TIMER1 is cleared at the start and the functions terminate if it exceeds
a MaxTime set at the start of each function.  If there is a problem with
getHourI2C, the old value of hour is returned by the function.