Fix Hourly Reporting
Added an update of prevHour = hour to the noonMessage routine.  Without
this, the system was sending the noon message and then sending it again
as an hourly debug message.  This messed up the resetting of EEPROM.

also removed the updatig of the internal RTCC within the noon message
routine since we don't use the internal RTCC.  This may need to be put
back when we move to using the internal RTCC