Code Installed Upon Leaving Ghana Aug 2017
When sleeping, only check the battery every 20minutes
Only set the time when the system is turning on for the first time
Use batteryLevelConstant which can be unique for each board (Vcc and
resistor divider)
Only calculate Leak Rate if someone has pumped at least 10liters.  This
is to avoid seeing a high leak rate if someone stops just as the water
splashes up over the sensor and then back down.