Add Enable Diagnostic messages capability
Add the AW_D: 0 or 1 message to be sent to the FONA.  If it is a 0,
disable the hourly diagnostic messages.  If it is a 1 enable them.

changed the main code that looked for the signal strength so that it
would not hang waiting for the network.  It now uses TRYtoconect to
network rather than waiting until network is found.

Added a reset of watchdog timer to the TryToConnectTonetwork routine
since we spend up to 45 sec there and the calling routie may not
anticipate that long of a delay

cleaned code up to get rid of some WARNINGS