Added verification messages and EEPROM phone number for diagnostic
Still need EEPROM write for when a text message turns on diagnostic message or changes phone number for either diagnostic or noon message. needs functionality to send the verification message when appropriate. *Before creating messages, make sure that the FONA has network connection or the wireless signal will be zero