Hourly Message Checking & Change Main Phone Number
Changed the routine which checks SMS incoming messages (Was
Installationmessages now CheckIncommingTextMessages) so that it is used
both when the diagnostic board is plugged in and each hour.

Added ChangeDailyReportPhoneNumber() so that an AWPN message allows the
user to change the phone number used for daily report

Simplified use of a variable FONAisON to keep track of whether the SIM
is on or off.