Sleep Decision
Changed the sleep decision to occur when the battery voltage starts
changing at a rate defined by BatteryDyingThreshold.  When the change is
more than this over a two hour period, go to sleep.  We still use 3.2 as
a hard stop.  Within the sleep routine, update the VTCC when you wake up
every 24 min so it is reasonably accurate when we wake up.  Since this
is also done every second, added the function updateVTCC to be used in
normal time keeping and when sleeping.