Add EEPROM storage of Main and Diagnostic Phone Numbers
Made the Change Time function treat hours and minutes together so that,
for example, +50 min will increase the hour if the current minute is
more than 9.

Added functions to save phone numbers (or country code) to EEPROM and to
read EEPROM back into a phone number string.

Added the saving of changed debug and Main phone numbers to EEPROM so
they are still correct after a reset (not original program)

Added a diagnostic which should be removed later.  The system sends a
text message to the diagnostic phone number when it first enters SLEEP
mode and again when it leaves it.  The messages include the values in
the BatteryVoltage array.