Change the ReadWaterSensor
Now that we have three states (Water, No Water and Broken) I changed the encoding so that 0 = Water, 1 = No Water and 2 = Broken Wire.  With this change, anything other than 0 means No Water (since we can't see water when the wire is broken we also interpret this as No Water).

Changed all routines which use the result of ReadWaterSensor to agree with the new output of the function