Renumber EEProm and Change Leak Rate and Prime reporting
Change report of Max Prime and Leak rate so that it runs from midnight to midnight like the volume rather than noon to noon.  This required the addition of 2 EEProm slots to differentiate between the current max leak and prime and the one from the previous day being held for the daily report.
To avoid errors, all references to EEProm locations were changed from numbers to variable names defined in a table in IWPUtilities